Update.
	* sunrpc/xdr_rec.c (xdrrec_getpos): Add cast to long to avoid warning.
	* sunrpc/xdr_sizeof.c (x_inline): Likewise.

+++ sunrpc/xdr_rec.c
@@ -304,7 +304,7 @@ xdrrec_getpos (const XDR *xdrs)
   RECSTREAM *rstrm = (RECSTREAM *) xdrs->x_private;
   long pos;
 
-  pos = __lseek ((int) rstrm->tcp_handle, (long) 0, 1);
+  pos = __lseek ((int) (long) rstrm->tcp_handle, (long) 0, 1);
   if (pos != -1)
     switch (xdrs->x_op)
       {
